вроде как финиш

+++ admin/controllers/DeleteControllers.php
@@ -2,6 +2,7 @@
 
 namespace admin\controllers;
 
+use admin\controllers\log\Logger;
 use admin\models\Article;
 use Exception;
 
@@ -11,12 +12,21 @@ class DeleteControllers extends Controllers
      * класс-контроллер
      * удаляет новость и редиректит на главную страницы
      * @return void
+     * @throws \Exception
      */
-    protected function handle():void
+    protected function handle(): void
     {
-        $article = new Article();
-        $article->id = $_POST['delete'];
-        $article->delete();
-        header('Location: ?ctrl=index');
+        if (!empty($_POST['delete'])) {
+            $article = new Article();
+            $article->id = $_POST['delete'];
+            $article->delete();
+            header('Location: ?ctrl=index');
+        } else {
+            $log = new Logger();
+            $log->loog(new Exception('что то пошло не так, не получилось удалить'));
+            throw new Exception('что то пошло не так, не получилось удалить');
+        }
+
+
     }
 }
